logger was added

+++ src/main/java/Counter.java
@@ -1,11 +1,11 @@
 public class Counter {
-    private int count;
+    private long count;
 
-    public int getCount() {
+    public long getCount() {
         return count;
     }
 
-    public void setCount(int count) {
+    public void setCount(long count) {
         this.count = count;
     }
 }
